android: Include git revision in mesa version
Running git_sha1_gen.py throws an exception and leaves revision empty
because .git directory is not copied to the build directory. Copying
.git to the build is not necessary and in Android's case it also contains
symbolic links with relative paths to repo directory which would make
the copying a bit tricky to keep a working .git directory.

Workaround this by using MESA_GIT_SHA1_OVERRIDE environment variable.
This fails if mesa doesn't have .git directory which is usually not
the case with Android builds.

Change-Id: I29d3cae05ff84fab6171731fadfb0e7d54e9e3e7