Add *_dri.so symlinks to Android build recipe
Android changed so that every file ending up on the target needs a build
rule. This is also true for symlinks. Until now the *_dri.so symlinks
where generated in meson and where unconditionally copied into the
target image. Since 152cdfab7c2 in Android build/make this doesn't
happen anymore and needs fixing in mesa.

Signed-off-by: Christian Poetzsch <christian.poetzsch@kernkonzept.com>
Tested-by: Konsta <konsta09@gmail.com>