android: Define a phony package that includes all defined mesa3d libs as dependency
Normally, all defined libs are needed for a target that uses mesa3d
and we have to manually add them to `PRODUCT_PACKAGES`. There
shouldn't be a module that's defined but unneeded.

After this change, we could build all of them automatically just by
including `mesa3d` in `PRODUCT_PACKAGES`.